Move integration tests to a subdirectory
